--- Grubitecht/Assets/Scripts/Objects/GroundTile.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e518c8b9d717dc243a2bd63f7b84af50
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
