+++ Grubitecht/Assets/Scripts/Objects/Objective.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b66a08841d9835743b267a2c76463113
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
