--- Grubitecht/Assets/Scripts/Pathfinding.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0d5b734328431a641ba4bf9e2a3e5593
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
