+++ Grubitecht/Assets/Scripts/World.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bcdaac4f1d5987f4583dfde731195b24
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
