+++ Grubitecht/Assets/Scripts/World/GridNavigator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3521fcf41e3085f43b22409430828c5e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
