+++ Grubitecht/Assets/Scripts/World/Pathfinder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 67169e17f0d81c04ebe0be26597674c2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
